refactor: News, NewsItem components
